pi_cam: added cropping
pi_cam: Added "basic" type checking for vscode

+++ Code/pi_cam/imageService.py
@@ -1,5 +1,6 @@
 import codecs
 import threading
+import typing
 import numpy as np
 import datetime
 import math
@@ -21,9 +22,10 @@ import ctypes
 from PIL import Image
 
 class ImageService:
-    def __init__(self, config: ServiceConfig, test_only=False):
+    def __init__(self, config: ServiceConfig, test_only=False, test_only_save_image=False):
         self.config = config
         self.test_only = test_only
+        self.test_only_save_image = test_only_save_image
         self.log = config.logger
         self.last_grab = 0
         bc = config.parser["breathecam"]
@@ -57,7 +59,7 @@ class ImageService:
                 img.mode = "RGBX"
             # Make up some extra EXIF data.
             zero_ifd = {piexif.ImageIFD.Make: "Raspberry Pi",
-                        piexif.ImageIFD.Model: self.picam2.camera.id,
+                        piexif.ImageIFD.Model: self.picam2.camera.id, # type: ignore
                         piexif.ImageIFD.Software: "Picamera2",
                         piexif.ImageIFD.MakerNoteSafety: 1}
             total_gain = metadata["AnalogueGain"] * metadata["DigitalGain"]
@@ -80,12 +82,19 @@ class ImageService:
         metadata = request.get_metadata() # Copies data from request
 
         # Create PIL image from request
-        img: Image = request.make_image("main") # Referenced data from request;  do not release request until after done with image
+        img = typing.cast(Image.Image, request.make_image("main")) # Referenced data from request;  do not release request until after done with image
         # Rotate image clockwise 90 degrees using numpy
 
         if rotate_ccw_90:
             img = img.rotate(90, expand=True)
 
+        if self.config.crop_top() or self.config.crop_bottom() or self.config.crop_left() or self.config.crop_right():
+            img = img.crop((
+                self.config.crop_left(), 
+                self.config.crop_top(), 
+                img.width - self.config.crop_right(), 
+                img.height - self.config.crop_bottom()
+                ))
 
         tmp_filename = f"{current_dir}/{capture_timestamp:.0f}-{os.getpid()}-tmp.jpg"
 
@@ -99,7 +108,12 @@ class ImageService:
         self.save_image(img, metadata, current_tmp_filename, format)
 
         if self.test_only:
-            os.unlink(current_tmp_filename)
+            if self.test_only_save_image:
+                os.rename(current_tmp_filename, "/tmp/test.jpg")
+                self.log.info("Saved test image to /tmp/test.jpg")
+            else:
+                os.unlink(current_tmp_filename)
+                self.log.info("Deleted test image")
             sys.exit(0)
 
         # Atomically replace current.jpg, for interactive display (e.g. focusing)
@@ -126,13 +140,13 @@ class ImageService:
 
         # Round down to nearest 180 for transform since transform can only do 0 or 180
         transform_rotation = rotate_ccw // 180 * 180
-        transform = libcamera.Transform(rotation=transform_rotation)
+        transform = libcamera.Transform(rotation=transform_rotation) # type: ignore
         rotate_ccw_90 = (rotate_ccw - transform_rotation) == 90
 
         # preview defaults to lower resolution, auto-exposure and auto-white-balance
         preview_config = self.picam2.create_preview_configuration()
         preview_config["transform"] = transform
-        self.picam2.configure(preview_config)
+        self.picam2.configure(preview_config) # type: ignore
 
         self.picam2.start()
         self.log.info("Running for 2 seconds in preview move to lock auto exposure")
@@ -146,7 +160,7 @@ class ImageService:
         self.log.info(f"Initial config is {still_config}")
         self.picam2.align_configuration(still_config)
         self.log.info(f"Config after alignment is {still_config}")
-        self.picam2.configure(still_config)
+        self.picam2.configure(still_config) # type: ignore
 
         self.picam2.start()
         logging.getLogger('picamera2').setLevel(logging.WARNING)
@@ -160,7 +174,7 @@ class ImageService:
         while True:
             # Capture next frame
             before = time.monotonic()
-            request: CompletedRequest = self.picam2.capture_request()
+            request: CompletedRequest = self.picam2.capture_request() # type: ignore
             #request.
             capture_duration = time.monotonic() - before
             self.log.debug(f"Captured frame in {capture_duration*1000:.1f}ms")
@@ -195,8 +209,13 @@ if __name__ == '__main__':
     os.chdir(script_dir)
     import argparse
     parser = argparse.ArgumentParser()
-    parser.add_argument("--test-only", action="store_true", help="Test image capture and exit")
+    parser.add_argument("--test-only", action="store_true", help="Test image capture, delete image, and exit")
+    parser.add_argument("--test-only-save-image", action="store_true", help="Save image in /tmp during test")
+    parser.add_argument("--simulate-camera", action="store_true", help="Simulate camera capture without opening camera")
     args = parser.parse_args()
 
-    svc = ImageService(ServiceConfig('./', 'image'), test_only=args.test_only)
+    svc = ImageService(
+        ServiceConfig('./', 'image'),
+        test_only=args.test_only,
+        test_only_save_image=args.test_only_save_image)
     svc.run()
